+++ src/components/wistia-video/index.js
@@ -1,6 +1,7 @@
 import React, { useEffect } from 'react';
 import { useGlobals } from 'State/globals';
 import { useTracker } from 'State/segment-analytics';
+import useScript from 'Hooks/use-script';
 
 const WistiaVideo = React.forwardRef(({ videoId }, ref) => {
   const { location } = useGlobals();
@@ -9,32 +10,12 @@ const WistiaVideo = React.forwardRef(({ videoId }, ref) => {
     url: location.pathname,
   });
 
-  useEffect(() => {
-    const loadedScripts = new Set();
-    const scriptTags = document.querySelectorAll('script');
-    scriptTags.forEach((node) => {
-      if (node.src) {
-        // strip protocol so we can match //fast.wistia.com...
-        loadedScripts.add(node.src.split(location.protocol)[1]);
-      }
-    });
-
-    function loadScript(src) {
-      if (!loadedScripts.has(src)) {
-        const script = document.createElement('script');
-        script.src = src;
-        script.async = true;
-        document.head.appendChild(script);
-        loadedScripts.add(src);
-      }
-    }
-
-    // the server sends these script tags on pages that have wistia videos
-    // you can navigate to a page with a wistia video via an in-page link,
-    // so we need to check that the script tags are on this page, and load them if not
-    loadScript(`//fast.wistia.com/embed/medias/${videoId}.jsonp`);
-    loadScript('//fast.wistia.com/assets/external/E-v1.js');
-  }, []);
+  // the server writes these script tags on pages that have wistia videos,
+  // but you can navigate to a page with a wistia video via in-page link,
+  // (which won't be server-side rendered), so we need to append them to the
+  // document to ensure they have actually been loaded
+  useScript(`//fast.wistia.com/embed/medias/${videoId}.jsonp`);
+  useScript('//fast.wistia.com/assets/external/E-v1.js');
 
   useEffect(() => {
     /* eslint-disable no-underscore-dangle */
